pyproject.toml
[tool.poetry]
name = "travel-planner"
version = "0.1.0"
description = "One stop solution to travel planning"
authors = ["Atishay Jain, Naisheel Shah, Samarth Ghulyani, Ojas Naik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-search-results = "^2.4.2"
langchain = "^0.0.149"
openai = "^0.27.4"
uagents = "*"
google-generativeai = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
